@@ -1,12 +1,6 @@
-import Vue from 'vue'
+import { createApp } from 'vue'
 import App from './App.vue'
-import store from './store'
 import router from './router'
+import store from './store'
 
-Vue.config.productionTip = false
-
-new Vue({
-  store,
-  router,
-  render: h => h(App)
-}).$mount('#app')
+createApp(App).use(store).use(router).mount('#app')
