@@ -1,48 +1,98 @@
-import Vue from 'vue'
-import VueRouter from 'vue-router'
-import UtilOneComponentVue from "@/components/utilOne.vue";
-import roadingComponentVue from "@/components/roadingComponent.vue";
-import resultComponentVue from "@/components/resultComponent.vue";
-import startComponentVue from "@/components/startComponent.vue"
-import watermarkComponent from "@/components/watermarkComponent.vue";
+import { createRouter, createWebHistory } from 'vue-router'
 
-Vue.use(VueRouter)
+import MainView from "@/views/MainView.vue";
+import loginView from "@/views/loginView.vue";
+
+import MyPageView from "@/views/MyPage/mainView.vue";
+import MyPageClosetView from "@/views/MyPage/ClosetView.vue";
+import MyPageCodybookView from "@/views/MyPage/CodybookView.vue";
+import MyPageStyleView from "@/views/MyPage/StyleView.vue";
+import AddClothesView from "@/views/AddClothesView.vue";
+
+import FindMainView from "@/views/Find/MainView.vue";
+import FindResultView from "@/views/Find/ResultView.vue";
+import FindErrorView from "@/views/Find/ErrorView.vue";
+
+import CodybookView from "@/views/CodybookView.vue";
+
+import RecomView from "@/views/RecomView.vue";
 
 const routes = [
   {
-    path: '/util',
-    name: 'UtilOneVue',
-    component: UtilOneComponentVue,
-    children: [
+    path: '/',
+    name: 'Main',
+    component: MainView
+  },
+  {
+    path: '/login',
+    name: 'login',
+    component:loginView 
+  },
+  {
+    path: '/MyPage',
+    children:[
       {
-        path:'start',
-        component: startComponentVue
+        path: '',
+        name: 'MyPageView',
+        component:MyPageView,
       },
       {
-        path:'roading',
-        component: roadingComponentVue,
+        path: 'style',
+        name: 'MyPageStyleView',
+        component: MyPageStyleView
       },
       {
-        path:'result',
-        component: resultComponentVue,
+        path: 'closet',
+        name: 'MyPageClosetView',
+        component: MyPageClosetView
+      },
+      {
+        path: 'codybook',
+        name: 'MyPageCodybookView',
+        component: MyPageCodybookView
       }
     ]
   },
   {
-    path: '/watermark',
-    name:'water',
-    component: watermarkComponent
+    path: '/AddClothes',
+    name: 'AddClothesView',
+    component: AddClothesView
   },
   {
-    path: '/modal',
-    name:'modal',
-    component: watermarkComponent
-  }
+    path: '/Find',
+    children:[
+      {
+        path:'',
+        name: 'FindMainView',
+        component: FindMainView
+      },
+      {
+        path: 'Result',
+        name: 'FindResultView',
+        component: FindResultView
+      },
+      {
+        path: 'Error',
+        name: 'FindErrorView',
+        component: FindErrorView
+      }
+    ]
+  },
+  {
+    path: '/Codybook',
+    name: 'CodybookView',
+    component: CodybookView
+  },
+  {
+    path: '/Recom',
+    name: 'RecomView',
+    component: RecomView
+  },
+  
 ]
 
-const router = new VueRouter({
-  mode: 'history',
-  base: process.env.BASE_URL,
+const router = createRouter({
+  history: createWebHistory(process.env.BASE_URL),
   routes
 })
 
